post review

+++ consensus/qbft/src/main/java/org/hyperledger/besu/consensus/qbft/payload/MessageFactory.java
@@ -27,6 +27,7 @@ import org.hyperledger.besu.ethereum.core.Block;
 import org.hyperledger.besu.ethereum.core.Hash;
 import org.hyperledger.besu.ethereum.core.Util;
 
+import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Optional;
@@ -95,7 +96,7 @@ public class MessageFactory {
               roundIdentifier,
               Optional.of(
                   new PreparedRoundMetadata(
-                      preparedRoundData.get().preparedBlock.getHash(), round)));
+                      preparedRoundData.get().preparedBlock.getHash(), BigInteger.valueOf(round))));
       return new RoundChange(
           createSignedMessage(payload),
           Optional.of(preparedRoundData.get().preparedBlock),
